requirements.txt
asgiref==3.9.0
async-timeout==5.0.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.10.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
whitenoise==6.9.0
